Wing Generation Prelims
- added 2 scriptable objects: 1 to represent a type of tile
and the other to hold a list of tiletypes
representing a palette of tiletypes, intended to
be centered graphicly around a single "theme"
- cleared out some of the world controller code
- created FullGridWingController to control wings that
generate and store the data for their "features" (i.e.
rooms, corridors, etc.) on a fullgrid basis (i.e. a 2d array
of tiles) with no 2 features overlaping
Note: FullGridWingController no where near
functioning at this point
- created FullGridTest class to modify the inspector GUI
for FullGridWingController
- created a temp Wing Prefab to hold the
FullGridWingController
- created a couple testing tiletypes